Adapting QAtrainsim macro for ROOT6
